+++ lib/base/utils/logger.dart
@@ -1,10 +1,16 @@
 //import 'package:colorize/colorize.dart';
 
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
 import 'package:g_base_package/base/app_exception.dart';
 
 import '../provider/instance_provider.dart';
 import 'package:flutter/foundation.dart' as Foundation;
 import 'package:logger/logger.dart';
+
+import 'files.dart';
 
 class Log {
   ///This tags will be set allays in logs as prefix for easy sorting only app
@@ -29,6 +35,15 @@ class Log {
     ),
   );
 
+  static init({File fileToLog, Future<File> Function(File currentFile) getNewFile}) {
+    _logger = Logger(
+        printer: PrettyPrinter(
+          printTime: false,
+          methodCount: 0,
+        ),
+        output: MultiOutput([ConsoleOutput(), CustomFileOutput(file: fileToLog, getNewFile: getNewFile)]));
+  }
+
   //TODO set live template
 
   ///This method will print developer's info in logs only if we are in debug
@@ -51,7 +66,7 @@ class Log {
     printInDebugOnly(tag != null ? '$tagWarning $appTag $tag' : '$tagWarning $appTag', log, Level.warning);
   }
 
-  static void printInDebugOnly(String tag, String log, Level level , {bool addToCrashReporter = true}) {
+  static void printInDebugOnly(String tag, String log, Level level, {bool addToCrashReporter = true}) {
     if (fromUI && addToCrashReporter) {
       InstanceProvider.getInstance()?.crashReporter?.log(log, tag); //always save in Crash Reporter
     }
@@ -89,9 +104,9 @@ class Log {
   }
 
   static _fixError(error) {
-    if(error == null){
-      error  = AppException(data: "Handled error!");
-    }else if (!(error is Error)) {
+    if (error == null) {
+      error = AppException(data: "Handled error!");
+    } else if (!(error is Error)) {
       error = AppException(data: error);
     }
     return error;
@@ -121,5 +136,197 @@ class Log {
   ///Use this method to print in logs user's information messages.
   static i(String log, [String tag]) {
     printInDebugOnly(tag != null ? '$tagInfo $appTag $tag' : '$tagInfo $appTag', log, Level.info);
+  }
+}
+
+class FileLogs {
+  Future init({String dirName = 'logs', @required String fileName, bool deleteOtherLogFiles = true}) async {
+    Directory dir = await BaseFileUtils.getLocalDir(dirName);
+
+    var files = dir.listSync();
+
+    int latestVersion = 0;
+
+    String initialFileName = fileName;
+
+    for (final localFile in files) {
+      if (getFileName(localFile.path).startsWith(initialFileName)) {
+        // we have file for today
+        int version = getFileVersion(localFile.path);
+
+        if (version >= latestVersion) {
+          fileName = '${initialFileName}_${version + 1}';
+        }
+      } else if(deleteOtherLogFiles){
+        // this is logs from previous day delete it to free space
+        localFile.delete();
+      }
+    }
+
+    File fileToLog = await BaseFileUtils.getLocalFile(dirName, fileName);
+
+    fileToLog = await fileToLog.writeAsString(
+      "\n\n*******************************************"
+      "\nNew Session - ${DateTime.now().toIso8601String()}"
+      "\n*******************************************\n\n",
+      mode: FileMode.append,
+    );
+
+    Log.init(
+        fileToLog: fileToLog,
+        getNewFile: (currentFile) {
+          int version = getFileVersion(currentFile.path);
+
+          String fileName = '${initialFileName}_${version + 1}';
+
+          return BaseFileUtils.getLocalFile(dirName, fileName);
+        });
+  }
+
+  String getFileName(String path) {
+    int position = path.lastIndexOf("/");
+
+    if (position == -1) return path;
+
+    return path.substring(path.lastIndexOf("/") + 1);
+  }
+
+  int getFileVersion(String path) {
+    int position = path.lastIndexOf("_");
+
+    if (position == -1) return 0;
+
+    String version = path.substring(path.lastIndexOf("_") + 1);
+
+    return int.tryParse(version) ?? 0;
+  }
+
+  Future<List<File>> getLogFileVersions({String dirName = 'logs', @required String fileName}) async {
+    Directory dir = await BaseFileUtils.getLocalDir(dirName);
+
+    var files = dir.listSync();
+
+    String initialFileName = fileName;
+
+    List<File> fileVersions = [];
+
+    for (final localFile in files) {
+      if (getFileName(localFile.path).startsWith(initialFileName)) {
+        fileVersions.add(localFile);
+      }
+    }
+
+    return fileVersions;
+  }
+}
+
+class CustomFileOutput extends LogOutput {
+  File file;
+  final bool overrideExisting;
+  final Encoding encoding;
+  Future<File> Function(File currentFile) getNewFile;
+  IOSink _sink;
+  int linesCount = 0;
+  bool isOpeningNewFile = false;
+
+  CustomFileOutput({
+    @Foundation.required this.file,
+    this.overrideExisting = false,
+    this.encoding = utf8,
+    this.getNewFile,
+  });
+
+  @override
+  void init() {
+    linesCount = 0;
+
+    if (file != null) {
+      _sink = file.openWrite(
+        mode: overrideExisting ? FileMode.writeOnly : FileMode.writeOnlyAppend,
+        encoding: encoding,
+      );
+    }
+  }
+
+  @override
+  void output(OutputEvent event) {
+    if (isOpeningNewFile) return;
+
+    try {
+      _sink?.writeAll(event.lines, '\n');
+
+      linesCount += event?.lines?.length ?? 0;
+
+      //if lines are > X open new file
+      if (getNewFile != null && linesCount >= 10000) {
+        _openNewFile();
+      }
+    } catch (e) {
+      print(e);
+    }
+  }
+
+  @override
+  void destroy() async {
+    await _sink?.flush();
+    await _sink?.close();
+  }
+
+  Future _openNewFile() async {
+    isOpeningNewFile = true;
+
+    File newFile = await getNewFile(file);
+
+    if (newFile != null) {
+      await _sink?.flush();
+
+      await _sink?.close();
+
+      _sink = null;
+
+      file = newFile;
+
+      init();
+    }
+
+    isOpeningNewFile = false;
+  }
+}
+
+/// Logs simultaneously to multiple [LogOutput] outputs.
+class MultiOutput extends LogOutput {
+  List<LogOutput> _outputs;
+
+  MultiOutput(List<LogOutput> outputs) {
+    _outputs = _normalizeOutputs(outputs);
+  }
+
+  List<LogOutput> _normalizeOutputs(List<LogOutput> outputs) {
+    final normalizedOutputs = <LogOutput>[];
+
+    if (outputs != null) {
+      for (final output in outputs) {
+        if (output != null) {
+          normalizedOutputs.add(output);
+        }
+      }
+    }
+
+    return normalizedOutputs;
+  }
+
+  @override
+  void init() {
+    _outputs.forEach((o) => o.init());
+  }
+
+  @override
+  void output(OutputEvent event) {
+    _outputs.forEach((o) => o.output(event));
+  }
+
+  @override
+  void destroy() {
+    _outputs.forEach((o) => o.destroy());
   }
 }