+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.25.12] - Option to save logs into a file and print in console at the same time. 
+Files are split by 10 000 lines. 
+Get all files for today.
+Save in crash reporter last call and response separately   
+
 ## [1.25.11] - SingleScrollView handle rotations
 
 ## [1.25.10] - Message for Socketexception